Fix database creation issue, removing find(13) bug
http://stackoverflow.com/questions/6241803/cant-manage-to-put-data-into-my-database

It looks like symfony 2 helpers aren't creating two things that you need for the sqlite implementation to work correctly.

Firstly, in the parameters.ini for the project (tweak to suit):

 database_path=%kernel.root_dir%/config/MyDatabase.db
Secondly, in the config.yml file in the doctrine.dbal section, add the following:

 path: %database_path%
Then run:

php app/console doctrine:database:create
and it should generate the database for you in the app/config folder.

You will need to run:

php app/console doctrine:schema:create

fixing the controller to make it look for all questions in database. When you
hit index now in question controller it should render the form and from there
you can start creating questions.

In the .gitignore I'm ignoring as well the database file generated in the config.